Fixed bug where the wrong regularised loss was minimised
There were several sources for this bug. The FISTA iterations
did not divide the Lipschitz by two when computing the step
size. Another source for this bug was multiplying the groups
by their sqrt size twice instead of once when computing
regularisation strength. The final source for this bug was a
missing *not* in front of the check for whether to use
the old, wrong, regularisation strength.